+++ internal/service/users.go
@@ -16,21 +16,21 @@ func NewUserServiceImpl(repo repository.UserRepository, cfg *config.Config) *Use
 	return &UserServiceImpl{repo: repo, cfg: cfg}
 }
 
-func (s *UserServiceImpl) GetAllUsers(limit, offset int) ([]models.ReadUserDTO, error) {
+func (s *UserServiceImpl) GetAllUsers(limit, offset uint64) ([]models.ReadUserDTO, error) {
 	return s.repo.GetAllUsers(limit, offset)
 }
 
-func (s *UserServiceImpl) GetUserByID(id int) (*models.ReadUserDTO, error) {
+func (s *UserServiceImpl) GetUserByID(id uint64) (*models.ReadUserDTO, error) {
 	return s.repo.GetUserByID(id)
 }
 
-func (s *UserServiceImpl) UpdateUser(id int, user models.UpdateUserDTO) (*models.ReadUserDTO, error) {
+func (s *UserServiceImpl) UpdateUser(id uint64, user models.UpdateUserDTO) (*models.ReadUserDTO, error) {
 	if user.Password != "" {
 		user.PasswordHash = utils.HashPassword(user.Password)
 	}
 	return s.repo.UpdateUser(id, user)
 }
 
-func (s *UserServiceImpl) DeleteUser(id int) error {
+func (s *UserServiceImpl) DeleteUser(id uint64) error {
 	return s.repo.DeleteUser(id)
 }
